requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
email_validator==2.2.0
idna==3.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
password-validator==1.0
phonenumbers==8.13.44
pycryptodomex==3.20.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
